создает абстрактный класс. рефакторинг классов компонентов
