+++ src/components/task/task.js
@@ -1,24 +1,13 @@
 import {createTaskTemplate} from "./components/task";
-import {createElement} from "./../../tools/utils";
+import AbstractComponent from "../abstract-component";
 
-export default class Task {
+export default class Task extends AbstractComponent {
   constructor(task) {
+    super();
     this._task = task;
-    this._element = null;
   }
 
   getTemplate() {
     return createTaskTemplate(this._task);
-  }
-
-  getElement() {
-    if (!this._element) {
-      this._element = createElement(this.getTemplate());
-    }
-    return this._element;
-  }
-
-  removeElement() {
-    this._element = null;
   }
 }
