+++ src/main.js
@@ -2,7 +2,7 @@ import SiteMenu from "./components/menu/site-menu";
 import Board from "./components/board/board";
 import Filter from "./components/filter/filter";
 import {filters} from "./mock/filter";
-import {render} from "./tools/utils";
+import {render} from "./tools/utils/render";
 import {TASK_COUNT, Position} from "./tools/consts";
 import {renderBoard} from "./components/board/render-board";
 import {generateTasks} from "./mock/task";
