+++ src/tools/utils/render.js
@@ -1,0 +1,12 @@
+import {Position} from "../consts";
+
+export const render = (container, element, place) => {
+  switch (place) {
+    case Position.AFTERBEGIN:
+      container.prepend(element);
+      break;
+    case Position.BEFOREEND:
+      container.append(element);
+      break;
+  }
+};
